Checking user auth, moved icons

+++ news/views.py
@@ -117,6 +117,9 @@ def add_link(request):
     context = get_context(request)
     context['page_title'] += " - Add link"
 
+    if not request.user.is_authenticated:
+        return render(request, app_name / 'missing_rights.html', context)
+
     # if this is a POST request we need to process the form data
     if request.method == 'POST':
         method = "POST"
@@ -158,6 +161,9 @@ def import_links(request):
     summary_text = ""
     context = get_context(request)
     context['page_title'] += " - Import links"
+
+    if not request.user.is_authenticated:
+        return render(request, app_name / 'missing_rights.html', context)
 
     # if this is a POST request we need to process the form data
     if request.method == 'POST':
@@ -202,6 +208,9 @@ def remove_link(request, pk):
     context = get_context(request)
     context['page_title'] += " - Remove link"
 
+    if not request.user.is_authenticated:
+        return render(request, app_name / 'missing_rights.html', context)
+
     ft = LinkDataModel.objects.filter(id=pk)
     if ft.exists():
         ft.delete()
@@ -213,6 +222,9 @@ def remove_link(request, pk):
 def remove_all_links(request):
     context = get_context(request)
     context['page_title'] += " - Remove all links"
+
+    if not request.user.is_authenticated:
+        return render(request, app_name / 'missing_rights.html', context)
 
     ft = LinkDataModel.objects.all()
     if ft.exists():
@@ -243,6 +255,9 @@ from .prjconfig import Configuration
 def configuration(request):
     context = get_context(request)
     context['page_title'] += " - Configuration"
+
+    if not request.user.is_authenticated:
+        return render(request, app_name / 'missing_rights.html', context)
     
     c = Configuration.get_object()
     context['directory'] = c.directory
@@ -258,6 +273,9 @@ def edit_link(request, pk):
     context = get_context(request)
     context['page_title'] += " - Edit link"
     context['pk'] = pk
+
+    if not request.user.is_authenticated:
+        return render(request, app_name / 'missing_rights.html', context)
 
     ft = LinkDataModel.objects.filter(id=pk)
     if not ft.exists():
